+++ src/settings/projectSetting.ts
@@ -41,7 +41,7 @@ const setting = {
     //显示图标
     showIcon: false,
   },
-  //菜单权限模式 ROLE 前端固定角色  BACK 动态获取
-  permissionMode: 'ROLE',
+  //菜单权限模式 FIXED 前端固定路由  BACK 动态获取
+  permissionMode: 'FIXED',
 };
 export default setting;